#212 Complete overhaul of sitemaps.
